requirements.txt
aws-psycopg2==1.2.1
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
DateTime==4.3
idna==2.10
kiwisolver==1.3.1
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
zope.interface==5.2.0
